correct navigation errors
